fix: train and ux and ztemp
